allow 6 digit shop price

+++ src/item.c
@@ -633,7 +633,7 @@ const u8 * ItemId_GetName(u16 itemId)
     return gItemsInfo[SanitizeItemId(itemId)].name;
 }
 
-u16 ItemId_GetPrice(u16 itemId)
+u32 ItemId_GetPrice(u16 itemId)
 {
     return gItemsInfo[SanitizeItemId(itemId)].price;
 }
